refactor(reactions): share MessageReactions across channel and DM
Moves MessageReactions.tsx from the channel page-view folder to
components/shared so the DM page can render the same chip + tooltip + emoji
picker without duplication. The component is callback-driven, so no source
prop is needed — the parent owns the API endpoint.

Updates channel page, ChannelView, and the channel index re-export to point
at the new location.
