+++ apps/web/src/components/layout/middle-content/page-views/channel/index.ts
@@ -1,4 +1,4 @@
 export { default as ChannelView } from './ChannelView';
 export { ChannelInput, type ChannelInputRef, type ChannelInputProps } from './ChannelInput';
 export { ChannelInputFooter, type ChannelInputFooterProps } from './ChannelInputFooter';
-export { MessageReactions, type Reaction, type MessageReactionsProps } from './MessageReactions';
+export { MessageReactions, type Reaction, type MessageReactionsProps } from '@/components/shared/MessageReactions';
